[Decomp][Unit] Cohesion split: extract damage mitigation (armor/absorb/resist) to UnitDamage.cpp
